:ok_hand:REFACTOR: Refatora a tela principal!
Refatora a tela principal, retirando os métodos e as variáveis
relacionadas ao processo de gerar o arquivo pdf. Além disso,
foi gerada uma nova classe na camada Controller, nomeada como
'ClasseRelatorio.java', esta sendo responsável por gerar o
documento pdf
